+++ tests/pytest_tests/unit_tests/test_util.py
@@ -1,4 +1,3 @@
-import datetime
 import os
 import platform
 import random
@@ -12,7 +11,6 @@ import matplotlib.pyplot as plt
 import numpy as np
 import plotly
 import pytest
-import requests
 import tensorflow as tf
 import wandb
 from wandb import util
@@ -472,152 +470,6 @@ def test_parse_tfjob_config():
     with mock.patch.dict("os.environ", {"TF_CONFIG": "LOL"}):
         assert util.parse_tfjob_config() is False
     assert util.parse_tfjob_config() is False
-
-
-###############################################################################
-# Test retry utilities
-###############################################################################
-
-
-def test_no_retry_auth():
-    e = mock.MagicMock(spec=requests.HTTPError)
-    e.response = mock.MagicMock(spec=requests.Response)
-    for status_code in (400, 409):
-        e.response.status_code = status_code
-        assert not util.no_retry_auth(e)
-    e.response.status_code = 401
-    e.response.reason = "Unauthorized"
-    with pytest.raises(wandb.CommError):
-        util.no_retry_auth(e)
-    e.response.status_code = 403
-    e.response.reason = "Forbidden"
-    with mock.patch("wandb.run", mock.MagicMock()):
-        with pytest.raises(wandb.CommError):
-            util.no_retry_auth(e)
-    e.response.status_code = 404
-    with pytest.raises(wandb.CommError):
-        util.no_retry_auth(e)
-
-    e.response = None
-    assert util.no_retry_auth(e)
-    e = ValueError("foo")
-    assert util.no_retry_auth(e)
-
-
-def test_check_retry_conflict():
-    e = mock.MagicMock(spec=requests.HTTPError)
-    e.response = mock.MagicMock(spec=requests.Response)
-
-    e.response.status_code = 400
-    assert util.check_retry_conflict(e) is None
-
-    e.response.status_code = 500
-    assert util.check_retry_conflict(e) is None
-
-    e.response.status_code = 409
-    assert util.check_retry_conflict(e) is True
-
-
-def test_check_retry_conflict_or_gone():
-    e = mock.MagicMock(spec=requests.HTTPError)
-    e.response = mock.MagicMock(spec=requests.Response)
-
-    e.response.status_code = 400
-    assert util.check_retry_conflict_or_gone(e) is None
-
-    e.response.status_code = 410
-    assert util.check_retry_conflict_or_gone(e) is False
-
-    e.response.status_code = 500
-    assert util.check_retry_conflict_or_gone(e) is None
-
-    e.response.status_code = 409
-    assert util.check_retry_conflict_or_gone(e) is True
-
-
-def test_make_check_reply_fn_timeout():
-    """Verify case where secondary check returns a new timeout."""
-    e = mock.MagicMock(spec=requests.HTTPError)
-    e.response = mock.MagicMock(spec=requests.Response)
-
-    check_retry_fn = util.make_check_retry_fn(
-        check_fn=util.check_retry_conflict_or_gone,
-        check_timedelta=datetime.timedelta(minutes=3),
-        fallback_retry_fn=util.no_retry_auth,
-    )
-
-    e.response.status_code = 400
-    check = check_retry_fn(e)
-    assert check is False
-
-    e.response.status_code = 410
-    check = check_retry_fn(e)
-    assert check is False
-
-    e.response.status_code = 500
-    check = check_retry_fn(e)
-    assert check is True
-
-    e.response.status_code = 409
-    check = check_retry_fn(e)
-    assert check
-    assert check == datetime.timedelta(minutes=3)
-
-
-def test_make_check_reply_fn_false():
-    """Verify case where secondary check forces no retry."""
-    e = mock.MagicMock(spec=requests.HTTPError)
-    e.response = mock.MagicMock(spec=requests.Response)
-
-    def is_special(e):
-        if e.response.status_code == 500:
-            return False
-        return None
-
-    check_retry_fn = util.make_check_retry_fn(
-        check_fn=is_special,
-        fallback_retry_fn=util.no_retry_auth,
-    )
-
-    e.response.status_code = 400
-    check = check_retry_fn(e)
-    assert check is False
-
-    e.response.status_code = 500
-    check = check_retry_fn(e)
-    assert check is False
-
-    e.response.status_code = 409
-    check = check_retry_fn(e)
-    assert check is False
-
-
-def test_make_check_reply_fn_true():
-    """Verify case where secondary check allows retry."""
-    e = mock.MagicMock(spec=requests.HTTPError)
-    e.response = mock.MagicMock(spec=requests.Response)
-
-    def is_special(e):
-        if e.response.status_code == 400:
-            return True
-        return None
-
-    check_retry_fn = util.make_check_retry_fn(
-        check_fn=is_special,
-        fallback_retry_fn=util.no_retry_auth,
-    )
-
-    e.response.status_code = 400
-    check = check_retry_fn(e)
-    assert check is True
-
-    e.response.status_code = 500
-    check = check_retry_fn(e)
-    assert check is True
-
-    e.response.status_code = 409
-    check = check_retry_fn(e)
-    assert check is False
 
 
 def test_downsample():
